refactor: 엔티티의 일급컬렉션 클래스 디렉토리 위치 변경
