+++ src/main/java/space/space_spring/domain/chat/chatroom/service/component/ChatRoomService.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import space.space_spring.domain.chat.chatroom.model.ChatRoom;
-import space.space_spring.domain.chat.chatroom.model.dto.ChatRooms;
-import space.space_spring.domain.chat.chatroom.model.dto.UserChatRooms;
+import space.space_spring.domain.chat.chatroom.model.ChatRooms;
+import space.space_spring.domain.chat.chatroom.model.UserChatRooms;
 import space.space_spring.domain.chat.chatroom.model.response.ChatRoomResponse;
 import space.space_spring.domain.chat.chatroom.model.response.ChatSuccessResponse;
 import space.space_spring.domain.chat.chatroom.model.response.CreateChatRoomResponse;
